# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Prueba tecnica
  version: 1.0.0
...
paths:
...
  /precios:
    get:
      tags:
        - Precios
      summary: Recupera todos los precios disponibles
      description: Recupera todos los precios disponibles en el sistema
      operationId: findAllPrecios
      security:
        - basicAuth: []
      responses:
        '200':
          description: Operacion con exito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrecioResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
    post:
      tags:
        - Precios
      summary: Inserta un precio en el sistema
      description: Inserta un precio en el sistema
      operationId: insertPrecio
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecioRequest'
      responses:
        '200':
          description: Operacion con exito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecioResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
    put:
      tags:
        - Precios
      summary: Actualiza un precio en el sistema
      description: Actualiza un precio en el sistema
      operationId: updatePrecio
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecioRequest'
      responses:
        '200':
          description: Operacion con exito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecioResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Not found
        '500':
          description: Internal server error
  /precios/{id}:
    delete:
      tags:
        - Precios
      summary: Borra un precio en el sistema
      description: Borra un precio en el sistema
      operationId: deletePrecio
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: id del precio a borrar
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Operacion con exito
        '400':
          description: Id del precio no valido
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
...
components:
  schemas:
    PrecioRequest:
      type: object
      properties:
        id:
          type: integer
          format: int32
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        precio:
          type: number
          format : float
        coche:
          $ref: '#/components/schemas/CocheRequest'
    PrecioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        precio:
          type: number
          format : float
    MarcaRequest:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nombre:
          type: string
...
    CocheRequest:
      type: object
      properties:
        id:
          type: integer
          format: int32
        modelo:
          type: string
        color:
          type: string
        cilindrada:
          type: integer
          format: int32
        potencia:
          type: integer
          format: int32
        marca:
          $ref: '#/components/schemas/MarcaRequest'
...
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
  responses:
    UnauthorizedError:
      description: Autenticacion incorrecta
      headers:
        WWW_Authenticate:
          schema:
            type: string